Move `modlogUtils.js` to new `moderation/` category
